Add symbols txt file for download feature.

+++ html/source/src/app/air/earnings-calendar/ecal-after-v2.sh
@@ -16,6 +16,7 @@ function ecalAfter {
 # Iterate through symbols and request after market price.
 # NOTE ecal update needs to run through 15 min after market close to get the close price off the delayed data
 echo '[' > data.json
+touch symbols.txt
 cat $ecalPath"ecal-update-symbols" | while read line || [ -n "$line" ]
 do
     # Define variables for symbol card and vitals
@@ -47,7 +48,6 @@ do
     # If the last tick price was later (in epoch time) than the close of regular hours, it's a after-market tick.
     if [[ $afterTimestamp -gt $regularHoursEpoch ]] ; then
         if (( $(echo "$afterPrice > $price" | bc -l) )) ; then
-echo 'made it'
             # Calculate percent change
             change=$(echo $afterPrice - $price | bc | awk '{printf "%f", $0}')
             percentCalc=$(echo $change / $price | bc -l)
@@ -76,6 +76,8 @@ echo 'made it'
             oneYearNull=$(./jq-linux64 '.[] | select(.symbol == "'$symbol'") | .oneYearNull' $dePath"data/ecal-intraday-data.json")
             # Build JSON
             echo '{"list":"After Market Movers","nullFlag":"","symbol": "'$symbol'","name": "'$name'","price": '$afterPrice',"dollarChange": '$change',"percentChange": '$changePercent',"time":'$time',"oneDayNull":'$oneDayNull',"oneDay": "http://localhost/source/src/app/air/earnings-calendar/data/charts/'$symbol'-1d.php","oneMonth": "http://localhost/source/src/app/air/earnings-calendar/data/charts/'$symbol'-1mo.php","threeMonthNull":'$threeMonthNull',"threeMonth": "http://localhost/source/src/app/air/earnings-calendar/data/charts/'$symbol'-3mo.php","sixMonthNull":'$sixMonthNull',"sixMonth": "http://localhost/source/src/app/air/earnings-calendar/data/charts/'$symbol'-6mo.php","oneYearNull":'$oneYearNull',"oneYear": "http://localhost/source/src/app/air/earnings-calendar/data/charts/'$symbol'-1yr.php","open": '$open',"high": '$high',"low":'$low',"volume": '$volume',"avgVol": '$avgVol',"sharesShort": '$sharesShort',"shortPercent": '$shortPercent',"marketCap": '$marketCap',"float": '$float',"headlines":'$headlines'},' >> data.json
+            # Copy symbol list for download
+            printf $symbol"\n" >> symbols.txt
         fi
     fi
 done
@@ -84,13 +86,16 @@ sed -i '$ s/.$//' data.json
 echo ']' >> data.json
 
 # Check for empty data set
-head -1 data.json | if [ "$1" == "" ]; then 
+wc=$(wc -l data.json | cut -d' ' -f1)
+if [ $wc -lt 5 ]; then
     rm data.json
+    rm symbols.txt
     ./ecal-v2.sh
 else
     # Clean up and prepare data for other scans
-    cp data.json /var/www/html/source/src/app/air/decision-engine/data/ecal-after-data.json
-    cat data.json > /var/www/html/source/src/app/air/earnings-calendar/data/data.json
+    cp data.json $dePath"ecal-after-data.json"
+    cat data.json > $ecalPath"data.json"
+    mv symbols.txt $ecalPath"symbols.txt"
 fi
 
 }
